% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcm-package.R
\docType{package}
\name{ghcm-package}
\alias{ghcm}
\alias{ghcm-package}
\title{ghcm: Functional Conditional Independence Testing with the GHCM}
\description{
To learn more about ghcm, start with the vignette:
`browseVignettes(package = "ghcm")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arlundborg/ghcm}
  \item Report bugs at \url{https://github.com/arlundborg/ghcm/issues}
}

}
\author{
\strong{Maintainer}: Anton Rask Lundborg \email{a.lundborg@statslab.cam.ac.uk}

Authors:
\itemize{
  \item Rajen D. Shah
  \item Jonas Peters
}

}
