% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcm-package.R
\docType{package}
\name{ghcm}
\alias{ghcm}
\title{ghcm: A package for Functional Conditional Independence Testing}
\description{
To learn more about ghcm, start with the vignette:
`browseVignettes(package = "ghcm")`
}
