% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{ghcm_class_constructor}
\alias{ghcm_class_constructor}
\title{Class constructor for the \code{ghcm} class.}
\usage{
ghcm_class_constructor(
  test_statistic,
  p,
  limiting_covariance,
  asymptotic_distribution,
  alpha
)
}
\arguments{
\item{test_statistic}{Desc}

\item{p}{Desc}

\item{limiting_covariance}{Desc}

\item{asymptotic_distribution}{Desc}

\item{alpha}{Desc}
}
\description{
Class constructor for the \code{ghcm} class.
}
