% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{ghcm_test}
\alias{ghcm_test}
\title{Perform the GHCM using the residuals from regressing X on Z (resid_f)
 and regressing Y on Z (resid_g).}
\usage{
ghcm_test(
  resid_X_on_Z,
  resid_Y_on_Z,
  X_is_multivariate = NULL,
  Y_is_multivariate = NULL,
  fpca_method = "fpca.sc",
  b = 10000,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{resid_X_on_Z, resid_Y_on_Z}{Numeric vectors or matrices. Residuals
when regressing X (Y) on Z with a suitable regression method.}

\item{X_is_multivariate, Y_is_multivariate}{Logcal scalars. If TRUE, treats
X (Y) as a multivariate random variable and not a functional random
 variable observed on a grid. Thus no FPCA expansion is performed on
 X (Y) if this is TRUE.}

\item{fpca_method}{String or function. If a string is given, will search
the refund package for a function with the given name. If a function is
given it must return a matrix with the same number of rows as the input
and the coordinates of the input in its FPCA basis as each row.}

\item{b}{Positive integer. The number of samples from the estimated
limiting distribution to estimate the p-value.}

\item{alpha}{Numeric in the unit interval. Significance level of the test.}

\item{...}{Additional arguments to be passed to the fpca_method.}
}
\description{
Desc
}
\examples{


}
