% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inner_product_matrix_splines}
\alias{inner_product_matrix_splines}
\title{Computes the matrix of L2 inner products of the splines given in list_of_splines
as produced by splines::interpSpline. The splines are assumed to be
functions on the interval [from, to].}
\usage{
inner_product_matrix_splines(list_of_splines, from, to)
}
\arguments{
\item{list_of_splines}{list of interpSpline objects.}

\item{from, to}{limits of integration.}
}
\value{
matrix of inner products.
}
\description{
Computes the matrix of L2 inner products of the splines given in list_of_splines
as produced by splines::interpSpline. The splines are assumed to be
functions on the interval [from, to].
}
