% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sim_brownian_motion}
\alias{sim_brownian_motion}
\title{Simulate a Brownian Motion}
\usage{
sim_brownian_motion(n, grid, sigma = 1)
}
\arguments{
\item{n}{Positive integer, the number of sample paths to simulate.}

\item{grid}{Numeric vector with consecutive values in the unit interval,
the grid points to evalute the sample paths.}

\item{sigma}{Optionally, a positive number specifying the square root of
the variance of the Brownian motions to simulate. Defaults to 1.}
}
\value{
Returns an \eqn{n} x \code{len(grid)} matrix where each row
is a sample path evaluated at the grid points.
}
\description{
\code{sim_brownian_motion} returns \eqn{n} simulated Brownian motion sample
paths evaluated on a grid.
}
\examples{
grid <- seq(0, 1, length.out = 100)
n <- 1000

X <- ghcm:::sim_brownian_motion(n, grid)

sigma <- 2

Y <- ghcm:::sim_brownian_motion(n, grid, sigma)

}
