% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{ghcm_test}
\alias{ghcm_test}
\title{Perform the GHCM using the residuals from regressing X on Z (resid_f)
 and regressing Y on Z (resid_g).}
\usage{
ghcm_test(
  resid_X_on_Z,
  resid_Y_on_Z,
  b = 10000,
  X_is_multivariate = FALSE,
  Y_is_multivariate = FALSE,
  fpca_method = NULL,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{resid_X_on_Z, resid_Y_on_Z}{This is a description. Bladebla}

\item{b}{This is a description}

\item{X_is_multivariate, Y_is_multivariate}{This is a description. Bladebla}

\item{fpca_method}{This is a description. Bladebla}

\item{alpha}{This is a description}

\item{...}{Additional arguments to be passed to the fpca_method.}
}
\description{
Desc
}
